+++ authenticator/src/test/java/com/amplifyframework/ui/authenticator/AuthenticatorViewModelTest.kt
@@ -20,6 +20,8 @@ import aws.smithy.kotlin.runtime.http.HttpException
 import com.amplifyframework.auth.AuthUserAttributeKey.email
 import com.amplifyframework.auth.AuthUserAttributeKey.emailVerified
 import com.amplifyframework.auth.MFAType
+import com.amplifyframework.auth.cognito.exceptions.service.LimitExceededException
+import com.amplifyframework.auth.exceptions.SessionExpiredException
 import com.amplifyframework.auth.exceptions.UnknownException
 import com.amplifyframework.auth.result.AuthResetPasswordResult
 import com.amplifyframework.auth.result.step.AuthNextResetPasswordStep
@@ -32,6 +34,7 @@ import com.amplifyframework.ui.authenticator.util.AmplifyResult.Error
 import com.amplifyframework.ui.authenticator.util.AmplifyResult.Success
 import com.amplifyframework.ui.authenticator.util.AuthConfigurationResult
 import com.amplifyframework.ui.authenticator.util.AuthProvider
+import com.amplifyframework.ui.authenticator.util.LimitExceededMessage
 import com.amplifyframework.ui.authenticator.util.NetworkErrorMessage
 import com.amplifyframework.ui.testing.CoroutineTestRule
 import io.kotest.matchers.shouldBe
@@ -130,6 +133,21 @@ class AuthenticatorViewModelTest {
     }
 
     @Test
+    fun `getCurrentUser error with session expired exception during start results in being signed out`() = runTest {
+        coEvery { authProvider.fetchAuthSession() } returns Success(mockAuthSession(isSignedIn = true))
+        coEvery { authProvider.getCurrentUser() } returns AmplifyResult.Error(SessionExpiredException())
+
+        viewModel.start(mockAuthenticatorConfiguration())
+        advanceUntilIdle()
+
+        coVerify(exactly = 1) {
+            authProvider.fetchAuthSession()
+            authProvider.getCurrentUser()
+            authProvider.signOut()
+        }
+    }
+
+    @Test
     fun `when already signed in during start the initial state should be signed in`() = runTest {
         coEvery { authProvider.fetchAuthSession() } returns Success(mockAuthSession(isSignedIn = true))
         coEvery { authProvider.getCurrentUser() } returns Success(mockAuthUser())
@@ -200,6 +218,73 @@ class AuthenticatorViewModelTest {
 
         viewModel.signIn("username", "password")
         viewModel.currentStep shouldBe AuthenticatorStep.SignInConfirmTotpCode
+    }
+
+    @Test
+    fun `SMS MFA Code next step shows the SignInConfirmMfa screen`() = runTest {
+        coEvery { authProvider.fetchAuthSession() } returns Success(mockAuthSession(isSignedIn = false))
+        coEvery { authProvider.signIn(any(), any()) } returns Success(
+            mockSignInResult(signInStep = AuthSignInStep.CONFIRM_SIGN_IN_WITH_SMS_MFA_CODE)
+        )
+
+        viewModel.start(mockAuthenticatorConfiguration(initialStep = AuthenticatorStep.SignIn))
+
+        viewModel.signIn("username", "password")
+        viewModel.currentStep shouldBe AuthenticatorStep.SignInConfirmMfa
+    }
+
+    @Test
+    fun `Custom Challenge next step shows the SignInConfirmCustomAuth screen`() = runTest {
+        coEvery { authProvider.fetchAuthSession() } returns Success(mockAuthSession(isSignedIn = false))
+        coEvery { authProvider.signIn(any(), any()) } returns Success(
+            mockSignInResult(signInStep = AuthSignInStep.CONFIRM_SIGN_IN_WITH_CUSTOM_CHALLENGE)
+        )
+
+        viewModel.start(mockAuthenticatorConfiguration(initialStep = AuthenticatorStep.SignIn))
+
+        viewModel.signIn("username", "password")
+        viewModel.currentStep shouldBe AuthenticatorStep.SignInConfirmCustomAuth
+    }
+
+    @Test
+    fun `New Password next step shows the SignInConfirmNewPassword screen`() = runTest {
+        coEvery { authProvider.fetchAuthSession() } returns Success(mockAuthSession(isSignedIn = false))
+        coEvery { authProvider.signIn(any(), any()) } returns Success(
+            mockSignInResult(signInStep = AuthSignInStep.CONFIRM_SIGN_IN_WITH_NEW_PASSWORD)
+        )
+
+        viewModel.start(mockAuthenticatorConfiguration(initialStep = AuthenticatorStep.SignIn))
+
+        viewModel.signIn("username", "password")
+        viewModel.currentStep shouldBe AuthenticatorStep.SignInConfirmNewPassword
+    }
+
+    @Test
+    fun `Confirm SignUp next step, get error from resendSignUpCode, stays in SignIn screen`() = runTest {
+        coEvery { authProvider.fetchAuthSession() } returns Success(mockAuthSession(isSignedIn = false))
+        coEvery { authProvider.signIn(any(), any()) } returns Success(
+            mockSignInResult(signInStep = AuthSignInStep.CONFIRM_SIGN_UP)
+        )
+        coEvery { authProvider.resendSignUpCode(any()) } returns AmplifyResult.Error(mockAuthException())
+
+        viewModel.start(mockAuthenticatorConfiguration(initialStep = AuthenticatorStep.SignIn))
+
+        viewModel.signIn("username", "password")
+        viewModel.currentStep shouldBe AuthenticatorStep.SignIn
+    }
+
+    @Test
+    fun `Confirm SignUp next step shows the SignUpConfirm screen`() = runTest {
+        coEvery { authProvider.fetchAuthSession() } returns Success(mockAuthSession(isSignedIn = false))
+        coEvery { authProvider.signIn(any(), any()) } returns Success(
+            mockSignInResult(signInStep = AuthSignInStep.CONFIRM_SIGN_UP)
+        )
+        coEvery { authProvider.resendSignUpCode(any()) } returns Success(mockk())
+
+        viewModel.start(mockAuthenticatorConfiguration(initialStep = AuthenticatorStep.SignIn))
+
+        viewModel.signIn("username", "password")
+        viewModel.currentStep shouldBe AuthenticatorStep.SignUpConfirm
     }
 
     @Test
@@ -473,6 +558,18 @@ class AuthenticatorViewModelTest {
         viewModel.resetPassword("username")
         viewModel.confirmResetPassword("username", "password", "code")
         viewModel.currentStep shouldBe AuthenticatorStep.SignIn
+    }
+
+    @Test
+    fun `Password reset results in limit exceeded message`() = runTest {
+        coEvery { authProvider.fetchAuthSession() } returns Success(mockAuthSession(isSignedIn = false))
+        coEvery { authProvider.resetPassword(any()) } returns Error(LimitExceededException(null))
+
+        viewModel.start(mockAuthenticatorConfiguration(initialStep = AuthenticatorStep.PasswordReset))
+
+        viewModel.shouldEmitMessage<LimitExceededMessage> {
+            viewModel.resetPassword("username")
+        }
     }
 //endregion
 //region helpers