+++ liveness/src/main/java/com/amplifyframework/ui/liveness/camera/LivenessCoordinator.kt
@@ -53,6 +53,7 @@ import java.util.concurrent.Executors
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 import kotlinx.coroutines.MainScope
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
 internal typealias OnMuxedSegment = (bytes: ByteArray, timestamp: Long) -> Unit
@@ -141,6 +142,16 @@ internal class LivenessCoordinator(
 
     init {
         startLivenessSession()
+        MainScope().launch {
+            delay(5_000)
+            if (!previewTextureView.hasReceivedUpdate) {
+                val faceLivenessException = FaceLivenessDetectionException(
+                    "The camera failed to open within the allowed time limit.",
+                    "Ensure the camera is available to use and that no other apps are using it."
+                )
+                processSessionError(faceLivenessException, true)
+            }
+        }
         MainScope().launch {
             getCameraProvider(context).apply {
                 if (lifecycleOwner.lifecycle.currentState != Lifecycle.State.DESTROYED) {